Move InputUID to coreComponents directory
