+++ admin/src/components/index.js
@@ -1,5 +1,4 @@
 export { default as Field } from './Field';
 export { default as Initializer } from './Initializer';
-export { default as InputUID } from './InputUID';
 export { default as ListViewTableCell } from './ListViewTableCell';
 export { default as PermalinkUID } from './PermalinkUID';
