+++ admin/src/coreComponents/index.js
@@ -1,0 +1,2 @@
+// The following components were cloned from Strapi's core files with minor mods.
+export { default as InputUID } from './InputUID';
